+++ src/botdriver/src/game.rs
@@ -4,7 +4,7 @@ pub trait Game {
     type Config;
 
     // returns game state and initial status
-    fn init(config: Config, player_ids: Vec<usize>) -> (Self, GameOutput);
+    fn init(config: Config, player_ids: Vec<u64>) -> (Self, GameOutput);
     // process player input and execute a game turn
     fn step(&mut self, responses: &Vec<Response>) -> GameOutput;
 }
@@ -24,13 +24,13 @@ pub enum GameStatus {
 
 #[derive(Debug)]
 pub struct Prompt {
-    pub player_id: usize,
+    pub player_id: u64,
     pub data: String,
 }
 
 #[derive(Debug)]
 pub struct Response {
-    pub player_id: usize,
+    pub player_id: u64,
     pub data: String,
 }
 // TODO Improve error type. error-chain maybe?
@@ -41,6 +41,6 @@ pub enum Outcome {
 }
 
 pub struct PlayerScore {
-    pub player_id: usize,
+    pub player_id: u64,
     pub score: f64,
 }
